# frontend/openapi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "EventHub API",
    "version": "v1"
  },
  "servers": [
    { "url": "http://localhost:5170" }
  ],
  "paths": {
...
    "/api/organizers/{userId}": {
      "get": {
        "tags": ["Organizers"],
        "operationId": "getOrganizerProfile",
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OrganizerProfileDto" }
              }
            }
          }
        }
      }
    },
    "/api/organizers/me": {
      "put": {
        "tags": ["Organizers"],
        "operationId": "updateOrganizerProfile",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateOrganizerProfileDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OrganizerProfileDto" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "OrganizerProfileDto": {
        "type": "object",
        "properties": {
          "userId": { "type": "integer" },
          "email": { "type": "string" },
          "fullName": { "type": "string" },
          "companyName": { "type": "string" },
          "taxNumber": { "type": "string", "nullable": true },
          "isApproved": { "type": "boolean" },
          "approvedAt": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "UpdateOrganizerProfileDto": {
        "type": "object",
        "required": ["companyName"],
        "properties": {
          "companyName": { "type": "string" },
          "taxNumber": { "type": "string", "nullable": true }
        }
      },
...
    }
  },
  "security": [{ "Bearer": [] }]
}